Refactor API to put all the parsing and generator methods on XtcePacketDefinition so users don't have to instantiate separate definition and parser objects Remove word_size kwarg from packet generator Update user docs Change module name from parseables to packets Change parse_CCSDSPacket to parse_ccsds_packet to match PEP style Change parse_ccsds_packet to instance method instead of static method Change skip_header_bits to skip_header_bytes
