+++ space_packet_parser/comparisons.py
@@ -8,7 +8,7 @@ import warnings
 import lxml.etree as ElementTree
 
 from space_packet_parser.exceptions import ComparisonError
-from space_packet_parser import parseables
+from space_packet_parser import packets
 
 
 # Common comparable mixin
@@ -68,13 +68,13 @@ class MatchCriteria(AttrComparable, metaclass=ABCMeta):
 
     @abstractmethod
     def evaluate(self,
-                 packet: parseables.CCSDSPacket,
+                 packet: packets.CCSDSPacket,
                  current_parsed_value: Optional[Union[int, float]] = None) -> bool:
         """Evaluate match criteria down to a boolean.
 
         Parameters
         ----------
-        packet : parseables.CCSDSPacket
+        packet : packets.CCSDSPacket
             Packet data used to evaluate truthyness of the match criteria.
         current_parsed_value : any, Optional
             Uncalibrated value that is currently being matched (e.g. as a candidate for calibration).
@@ -157,14 +157,14 @@ class Comparison(MatchCriteria):
         return cls(value, parameter_name, operator=operator, use_calibrated_value=use_calibrated_value)
 
     def evaluate(self,
-                 packet: parseables.CCSDSPacket,
+                 packet: packets.CCSDSPacket,
                  current_parsed_value: Optional[Union[int, float]] = None) -> bool:
         """Evaluate comparison down to a boolean. If the parameter to compare is not present in the parsed_data dict,
         we assume that we are comparing against the current raw value in current_parsed_value.
 
         Parameters
         ----------
-        packet : parseables.CCSDSPacket
+        packet : packets.CCSDSPacket
             Packet data used to evaluate truthyness of the match criteria.
         current_parsed_value : Union[int, float]
             Optional. Uncalibrated value that is currently a candidate for calibration and so has not yet been added
@@ -321,13 +321,13 @@ class Condition(MatchCriteria):
                          'See 3.4.3.4.2 of XTCE Green Book CCSDS 660.1-G-2')
 
     def evaluate(self,
-                 packet: parseables.CCSDSPacket,
+                 packet: packets.CCSDSPacket,
                  current_parsed_value: Optional[Union[int, float]] = None) -> bool:
         """Evaluate match criteria down to a boolean.
 
         Parameters
         ----------
-        packet : parseables.CCSDSPacket
+        packet : packets.CCSDSPacket
             Packet data used to evaluate truthyness of the match criteria.
         current_parsed_value : Optional[Union[int, float]]
             Current value being parsed. NOTE: This is currently ignored. See the TODO item below.
@@ -443,13 +443,13 @@ class BooleanExpression(MatchCriteria):
         raise ValueError(f"Failed to parse {element}")
 
     def evaluate(self,
-                 packet: parseables.CCSDSPacket,
+                 packet: packets.CCSDSPacket,
                  current_parsed_value: Optional[Union[int, float]] = None) -> bool:
         """Evaluate the criteria in the BooleanExpression down to a single boolean.
 
         Parameters
         ----------
-        packet : parseables.CCSDSPacket
+        packet : packets.CCSDSPacket
             Packet data used to evaluate truthyness of the match criteria.
         current_parsed_value : Optional[Union[int, float]]
             Current value being parsed.
@@ -531,12 +531,12 @@ class DiscreteLookup(AttrComparable):
 
         return cls(match_criteria, lookup_value)
 
-    def evaluate(self, packet: parseables.CCSDSPacket, current_parsed_value: Optional[Union[int, float]] = None) -> Any:
+    def evaluate(self, packet: packets.CCSDSPacket, current_parsed_value: Optional[Union[int, float]] = None) -> Any:
         """Evaluate the lookup to determine if it is valid.
 
         Parameters
         ----------
-        packet : parseables.CCSDSPacket
+        packet : packets.CCSDSPacket
             Packet data used to evaluate truthyness of the match criteria.
         current_parsed_value: Optional[Union[int, float]]
             If referenced parameter in criterion isn't in the packet, we assume we are comparing against this
