+++ tests/integration/test_bufferedreader_parsing.py
@@ -1,7 +1,7 @@
 """Integration test for parsing JPSS packets"""
 # Local
 from space_packet_parser import definitions
-from space_packet_parser import parser, parseables
+from space_packet_parser import packets
 
 
 def test_jpss_xtce_packet_parsing(jpss_test_data_dir):
@@ -9,15 +9,14 @@ def test_jpss_xtce_packet_parsing(jpss_test_data_dir):
     jpss_xtce = jpss_test_data_dir / 'jpss1_geolocation_xtce_v1.xml'
     jpss_definition = definitions.XtcePacketDefinition(xtce_document=jpss_xtce)
     assert isinstance(jpss_definition, definitions.XtcePacketDefinition)
-    jpss_parser = parser.PacketParser(jpss_definition)
 
     jpss_packet_file = jpss_test_data_dir / 'J01_G011_LZ_2021-04-09T00-00-00Z_V01.DAT1'
 
     with jpss_packet_file.open('rb') as binary_data:
-        jpss_packet_generator = jpss_parser.generator(binary_data, show_progress=True)
+        jpss_packet_generator = jpss_definition.packet_generator(binary_data, show_progress=True)
         n_packets = 0
         for jpss_packet in jpss_packet_generator:
-            assert isinstance(jpss_packet, parseables.CCSDSPacket)
+            assert isinstance(jpss_packet, packets.CCSDSPacket)
             assert jpss_packet.header['PKT_APID'].raw_value == 11
             assert jpss_packet.header['VERSION'].raw_value == 0
             n_packets += 1
