+++ tests/integration/test_xtce_based_parsing/test_ctim_parsing.py
@@ -1,6 +1,6 @@
 """Test parsing of CTIM instrument data"""
 # Local
-from space_packet_parser import definitions, parser
+from space_packet_parser import definitions
 
 
 def test_ctim_parsing(ctim_test_data_dir):
@@ -12,11 +12,10 @@ def test_ctim_parsing(ctim_test_data_dir):
 
     print("Loading and parsing data")
     test_packet_file = ctim_test_data_dir / 'ccsds_2021_155_14_39_51'
-    parser_inst = parser.PacketParser(pkt_def)
     with open(test_packet_file, 'rb') as pkt_file:
-        pkt_gen = parser_inst.generator(pkt_file,
-                                        root_container_name="CCSDSTelemetryPacket",
-                                        show_progress=True)
+        pkt_gen = pkt_def.packet_generator(pkt_file,
+                                           root_container_name="CCSDSTelemetryPacket",
+                                           show_progress=True)
         packets = list(pkt_gen)
 
     assert len(packets) == 1499
