+++ tests/integration/test_xtce_based_parsing/test_suda_parsing.py
@@ -5,7 +5,7 @@ The data used here is SUDA data but the fields are parsed using IDEX naming conv
 """
 # Local
 from space_packet_parser import definitions
-from space_packet_parser import parser, parseables
+from space_packet_parser import packets
 
 
 def parse_hg_waveform(waveform_raw: str):
@@ -47,21 +47,20 @@ def test_suda_xtce_packet_parsing(suda_test_data_dir):
     suda_xtce = suda_test_data_dir / 'suda_combined_science_definition.xml'
     suda_definition = definitions.XtcePacketDefinition(xtce_document=suda_xtce)
     assert isinstance(suda_definition, definitions.XtcePacketDefinition)
-    suda_parser = parser.PacketParser(suda_definition)
     suda_packet_file = suda_test_data_dir / 'sciData_2022_130_17_41_53.spl'
 
     with suda_packet_file.open('rb') as suda_binary_data:
-        suda_packet_generator = suda_parser.generator(suda_binary_data,
-                                                      skip_header_bits=32,
-                                                      show_progress=True)
+        suda_packet_generator = suda_definition.packet_generator(suda_binary_data,
+                                                                 skip_header_bytes=4,
+                                                                 show_progress=True)
         for suda_packet in suda_packet_generator:
-            assert isinstance(suda_packet, parseables.CCSDSPacket)
+            assert isinstance(suda_packet, packets.CCSDSPacket)
             assert suda_packet.header['PKT_APID'].raw_value == 1425, "APID is not as expected."
             assert suda_packet.header['VERSION'].raw_value == 0, "CCSDS header VERSION incorrect."
 
         suda_binary_data.pos = 0
-        suda_packet_generator = suda_parser.generator(suda_binary_data,
-                                                      skip_header_bits=32)
+        suda_packet_generator = suda_definition.packet_generator(suda_binary_data,
+                                                                 skip_header_bytes=4)
 
         try:
             p = next(suda_packet_generator)
